replace tyVarList with mkTemplateTyVars
tyVarList is not particularly safe to use when you need multiple TyVars
with different Kinds. The naive

  a:_ = tyVarList kind1
  b:_ = tyVarList kind2

is wrong because a and b will share a Unique and thus substitutions will
be applied incorrectly.

mkTemplateTyVars instead takes a list of kinds and creates the
appropriate number of TyVars, e.g.

  [a,b] = mkTemplateTyVars [kind1, kind2]