use `errorWithoutStackTrace` in base.
base does not export any API functions that use Implicit
CallStacks (apart from error and undefined), so the stack traces
produced by error calls in base tend to only contain the call to `error`
itself. This is not very helpful for users, as base already has a set of
descriptive error messages that explain which function failed, so the
stack trace is just noise.