don't infer CallStacks at the top-level.
Unless we're doing PartialTypeSignatures and one was provided. Note that
this means that

  foo :: _ => CallStack
  foo = getCallStack ?callStack

will be an *empty* CallStack, since we won't infer a CallStack for the
hole in the theta. I think this is the right move though, since we want
CallStacks to be opt-in. One can always write

  foo :: (HasCallStack, _) => CallStack
  foo = getCallStack ?callStack

to get the CallStack and still have GHC infer the rest of the theta.