Don't inline String literals or CallStacks
Summary:
This patch avoids inlining String literals and CallStacks by immediately floating them to top-level binders in the desugarer.

The only knock-on effect is that `DsGblEnv` has a new field to track extra top-level binders we may want to add. It's only used to track String literals and CallStacks at the moment. The new function `DsUtils.bindExprAtTopLevel` takes an `Expr` and attempts to create a new top-level binder for it (attempts, because we may be compiling a simple expression in which case the notion of a top-level binder makes no sense). If successsful it returns a `Var` that is bound to the `Expr`, otherwise it returns the `Expr` unmodified. The new function `DsMonad.withTopBinds` initializes the `DsGblEnv` to track additional top-level binders and returns any added binders alongside the result of the provided action.

`nofib` indicates a substantial reduction in both binary size and allocations with this patch.

```
--------------------------------------------------------------------------------
        Program           Size    Allocs   Runtime   Elapsed  TotalMem
--------------------------------------------------------------------------------
            Min         -15.6%    -10.2%     -4.1%    -13.4%    -20.0%
            Max          +1.3%     +0.8%     +7.1%     +9.3%     +6.7%
 Geometric Mean         -14.9%     -0.3%     +1.9%     +1.3%     -0.1%
```

Test Plan:
- validate (new test in `testsuite/tests/deSugar/should_compile/T10844.hs`)
- inspect nofib benchmarks to see if the binary sizes have gone down

Reviewers: nomeata, hvr, austin, bgamari, simonpj

Subscribers: simonmar, thomie

Differential Revision: https://phabricator.haskell.org/D1259

GHC Trac Issues: #10844